+++ docs/src/pages/beep.js
@@ -1,5 +1,6 @@
+/** @jsx jsx */
 import React from 'react'
-import { Box } from 'theme-ui'
+import { jsx } from 'theme-ui'
 import Logo from '../logo'
 import LogoWhite from '../logo/white'
 
@@ -7,8 +8,14 @@ export default props => (
   <>
     <LogoWhite />
     <Logo />
-    <Box p={4} mt={4} bg="black" color="white">
+    <div
+      sx={{
+        p: 4,
+        mt: 4,
+        bg: 'black',
+        color: 'white',
+      }}>
       <Logo />
-    </Box>
+    </div>
   </>
 )