+++ packages/core/src/index.js
@@ -29,6 +29,8 @@ export const createParser = config => {
   const cache = {}
   const parse = props => {
     let styles = {}
+    const isCacheDisabled = props.theme && props.theme.disableStyledSystemCache
+
     for (const key in props) {
       if (!config[key]) continue
       const sx = config[key]
@@ -37,10 +39,10 @@ export const createParser = config => {
 
       if (typeof raw === 'object') {
         cache.breakpoints =
-          cache.breakpoints ||
+          (!isCacheDisabled && cache.breakpoints) ||
           get(props.theme, 'breakpoints', defaults.breakpoints)
         if (Array.isArray(raw)) {
-          cache.media = cache.media || [
+          cache.media = (!isCacheDisabled && cache.media) || [
             null,
             ...cache.breakpoints.map(createMediaQuery),
           ]
@@ -67,6 +69,14 @@ export const createParser = config => {
   parse.config = config
   parse.propNames = Object.keys(config)
   parse.cache = cache
+
+  const keys = Object.keys(config).filter(k => k !== 'config')
+  if (keys.length > 1) {
+    keys.forEach(key => {
+      parse[key] = createParser({ [key]: config[key] })
+    })
+  }
+
   return parse
 }