+++ packages/css/test/index.js
@@ -253,3 +253,39 @@ test('skip breakpoints', () => {
     }
   })
 })
+
+test('padding shorthand does not collide with nested p selector', () => {
+  const result = css({
+    p: {
+      fontSize: 32,
+      color: 'tomato',
+      p: 2,
+    },
+    padding: 32,
+  })(theme)
+  expect(result).toEqual({
+    p: {
+      fontSize: 32,
+      color: 'tomato',
+      padding: 8,
+    },
+    padding: 32,
+  })
+})
+
+test('ignores array values longer than breakpoints', () => {
+  const result = css({
+    width: [ 32, 64, 128, 256, 512 ]
+  })({
+    breakpoints: [ '32em', '40em' ],
+  })
+  expect(result).toEqual({
+    width: 32,
+    '@media screen and (min-width: 32em)': {
+      width: 64,
+    },
+    '@media screen and (min-width: 40em)': {
+      width: 128,
+    },
+  })
+})